Somehow llList2Float never made it into LSL_BaseClass and LSL_BuiltIn_Commands_Interface - it is now (and foxes Mantis 395)

+++ OpenSim/Region/ScriptEngine/Common/LSL_BuiltIn_Commands.cs
@@ -1814,7 +1814,7 @@ namespace OpenSim.Region.ScriptEngine.Common
             {
                 return 0.0;
             }
-            return Convert.ToSingle(src.Data[index]);
+            return Convert.ToDouble(src.Data[index]);
         }
 
         public string llList2String(LSL_Types.list src, int index)
